Updating rule ID, class name and message
