+++ src/modeler/AutoRest.Swagger/Validation/ModelDescriptionMissing.cs
@@ -1,0 +1,48 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using AutoRest.Core.Logging;
+using AutoRest.Swagger.Validation.Core;
+using AutoRest.Core.Properties;
+using AutoRest.Swagger.Model;
+
+namespace AutoRest.Swagger.Validation
+{
+    public class ModelDescriptionMissing : TypedRule<Schema>
+    {
+        /// <summary>
+        /// Id of the Rule.
+        /// </summary>
+        public override string Id => "M4000";
+
+        /// <summary>
+        /// Violation category of the Rule.
+        /// </summary>
+        public override ValidationCategory ValidationCategory => ValidationCategory.SDKViolation;
+
+        /// <summary>
+        /// The template message for this Rule. 
+        /// </summary>
+        /// <remarks>
+        /// This may contain placeholders '{0}' for parameterized messages.
+        /// </remarks>
+        public override string MessageTemplate => Resources.ModelDescriptionMissing;
+
+        /// <summary>
+        /// The severity of this message (ie, debug/info/warning/error/fatal, etc)
+        /// </summary>
+        public override Category Severity => Category.Warning;
+
+        public override IEnumerable<ValidationMessage> GetValidationMessages(Schema schema, RuleContext context)
+        {
+            if (schema != null && string.IsNullOrEmpty(schema.Reference) && schema.RepresentsCompositeType())
+            {
+                if (schema.Description == null)
+                {
+                    yield return new ValidationMessage(new FileObjectPath(context.File, context.Path), this, context.Key);
+                }
+            }
+        }
+    }
+}
